Start to rewrite KVS in Go
The PingCap [talent plan][0] uses a key value store db as object of study
for a practical network application. The course is focused on the Rust
language. As an attempt to remove the rust of my GoLang skills (did
y'all see what I did here?) I'll rewrite part of that application
(original code is [here][1]).

This first commit creates the basic foundation for a `add` command to
`KVS`.

[0]: https://github.com/pingcap/talent-plan/tree/master/courses/rust
[1]: https://github.com/jonathanbeber/rust101/tree/master/talent-plan/pna